+++ src/test/java/io/github/acm19/aws/interceptorv5/test/Sample.java
@@ -11,51 +11,65 @@ package io.github.acm19.aws.interceptorv5.test;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.DefaultParser;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
+import org.apache.hc.client5.http.async.methods.SimpleHttpRequest;
+import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
+import org.apache.hc.client5.http.async.methods.SimpleRequestBuilder;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.apache.hc.client5.http.classic.methods.HttpPost;
 import org.apache.hc.client5.http.classic.methods.HttpUriRequest;
+import org.apache.hc.client5.http.config.TlsConfig;
 import org.apache.hc.client5.http.entity.UrlEncodedFormEntity;
+import org.apache.hc.client5.http.impl.async.CloseableHttpAsyncClient;
+import org.apache.hc.client5.http.impl.async.HttpAsyncClients;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
-import org.apache.hc.client5.http.impl.classic.CloseableHttpResponse;
 import org.apache.hc.client5.http.impl.classic.HttpClients;
+import org.apache.hc.client5.http.impl.nio.PoolingAsyncClientConnectionManagerBuilder;
+import org.apache.hc.core5.concurrent.FutureCallback;
+import org.apache.hc.core5.http.ContentType;
 import org.apache.hc.core5.http.HttpRequestInterceptor;
 import org.apache.hc.core5.http.HttpStatus;
 import org.apache.hc.core5.http.NameValuePair;
 import org.apache.hc.core5.http.message.BasicNameValuePair;
+import org.apache.hc.core5.http2.HttpVersionPolicy;
 import io.github.acm19.aws.interceptor.http.AwsRequestSigningApacheV5Interceptor;
 import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
 import software.amazon.awssdk.auth.signer.Aws4Signer;
 import software.amazon.awssdk.regions.Region;
 
 class Sample {
+    private static final int WAIT_FOR_RESPONSE = 5;
     private static final int SCREEN_WIDTH = 160;
 
     protected String endpoint;
     protected Region region;
     protected String service;
 
-    Sample(final String[] args) throws ParseException {
+    Sample(String[] args) throws ParseException {
         parseOptions(args);
     }
 
-    Sample(final String service, final String endpoint, final Region region) {
+    Sample(String service, String endpoint, Region region) {
         this.endpoint = endpoint;
         this.region = region;
         this.service = service;
     }
 
-    public static void main(final String[] args) throws IOException, ParseException {
+    public static void main(String[] args) throws IOException, ParseException {
         Sample sampleClass = new Sample(args);
         sampleClass.makeGetRequest();
         sampleClass.makePostRequest();
     }
 
-    private void parseOptions(final String[] args) throws ParseException {
+    private void parseOptions(String[] args) throws ParseException {
         Options options = new Options()
                 .addRequiredOption(null, "endpoint", true, "OpenSearch endpoint")
                 .addRequiredOption(null, "region", true, "AWS signing region")
@@ -99,20 +113,30 @@ class Sample {
         logRequest(httpPost);
     }
 
-    void logRequest(final HttpUriRequest request) throws IOException {
-        CloseableHttpClient httpClient = signingClient();
-        try (CloseableHttpResponse response = httpClient.execute(request)) {
-            switch (response.getCode()) {
-                case HttpStatus.SC_OK:
-                case HttpStatus.SC_CREATED:
-                    break;
-                default:
-                    throw new RuntimeException(response.getReasonPhrase());
-            }
+    private void makeAsynRequest() throws IOException {
+        String payload = "{\"aync\": \"request\"}";
+        SimpleHttpRequest request = SimpleRequestBuilder.post(endpoint + "/login")
+                .setBody(payload, ContentType.APPLICATION_JSON)
+                .build();
+
+        logAsyncRequest(request);
+    }
+
+    void logRequest(HttpUriRequest request) throws IOException {
+        try (CloseableHttpClient httpClient = signingClient()) {
+            httpClient.execute(request, response -> {
+                switch (response.getCode()) {
+                    case HttpStatus.SC_OK:
+                    case HttpStatus.SC_CREATED:
+                        return true;
+                    default:
+                        throw new RuntimeException(response.getReasonPhrase());
+                }
+            });
         }
     }
 
-    CloseableHttpClient signingClient() {
+    private CloseableHttpClient signingClient() {
         HttpRequestInterceptor interceptor = new AwsRequestSigningApacheV5Interceptor(
                 service,
                 Aws4Signer.create(),
@@ -121,6 +145,55 @@ class Sample {
 
         return HttpClients.custom()
                 .addRequestInterceptorLast(interceptor)
+                .build();
+    }
+
+    void logAsyncRequest(SimpleHttpRequest request) throws IOException {
+        try (CloseableHttpAsyncClient httpClient = signingAsyncClient()) {
+            httpClient.start();
+            Future<SimpleHttpResponse> response = httpClient.execute(request, new FutureCallback<SimpleHttpResponse>() {
+                @Override
+                public void completed(SimpleHttpResponse response) {
+                    switch (response.getCode()) {
+                        case HttpStatus.SC_OK:
+                        case HttpStatus.SC_CREATED:
+                            break;
+                        default:
+                            throw new RuntimeException(response.getReasonPhrase());
+                    }
+                }
+
+                @Override
+                public void failed(Exception ex) {
+                    throw new RuntimeException(ex);
+                }
+
+                @Override
+                public void cancelled() {
+                    throw new RuntimeException("Cancelled request.");
+                }
+            });
+
+            response.get(WAIT_FOR_RESPONSE, TimeUnit.MINUTES);
+        } catch (InterruptedException | ExecutionException | TimeoutException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    private CloseableHttpAsyncClient signingAsyncClient() {
+        HttpRequestInterceptor interceptor = new AwsRequestSigningApacheV5Interceptor(
+                service,
+                Aws4Signer.create(),
+                DefaultCredentialsProvider.create(),
+                region);
+
+        return HttpAsyncClients.custom()
+                .addRequestInterceptorLast(interceptor)
+                .setConnectionManager(PoolingAsyncClientConnectionManagerBuilder.create()
+                        .setDefaultTlsConfig(TlsConfig.custom()
+                                .setVersionPolicy(HttpVersionPolicy.FORCE_HTTP_1)
+                                .build())
+                        .build())
                 .build();
     }
 }